# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PUC HUB Student Portal API
  description: |
    Production-grade REST API for the Premier University Chittagong Student Portal.
    Authentication uses JWT access tokens (15-minute expiry) + httpOnly refresh token cookies (7-day expiry).
    All state-changing endpoints (POST/PUT/DELETE) require a valid `X-CSRF-Token` header.
  version: 1.1.0
  contact:
    name: PUC HUB Team
    email: support@puc-hub.vercel.app
servers:
  - url: https://puc-hub-api.onrender.com/api/v1
    description: Production (Render)
  - url: http://localhost:3000/api/v1
    description: Local Development
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using httpOnly cookie
      description: |
        Reads the `jid` cookie (set at login) and issues a new access token.
        Implements token rotation with reuse detection — a reused token invalidates the entire session.
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  accessToken:
                    type: string
        '401':
          description: Missing, expired, or invalid refresh token
        '403':
          description: Token reuse detected — session invalidated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...